handle time

+++ app/src/main/java/com/vad/ltale/data/User.kt
@@ -7,7 +7,7 @@ data class User(
 
     @SerializedName("id")
     @Expose
-    val userId: Int,
+    val userId: Long,
 
     @SerializedName("username")
     @Expose
